requirements.txt
beautifulsoup4==4.12.2
crispy-bootstrap5==0.7
django-crispy-forms==2.0
django-formtools==2.4
Django~=4.2.1
lxml==4.9.2
python-dotenv==1.0.0
httpx~=0.24.0